refactor(assembler): addr -> u8, immediate value -> u8

+++ crates/superr_assembler/src/ast.rs
@@ -1,5 +1,5 @@
-pub type Address = usize;
-pub type ImmediateValue = u64;
+pub type Address = u8;
+pub type ImmediateValue = u8;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum Operand {
